dev/experimental-utils/webcat/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "webcat"
version = "0.1.0"
description = "A CLI web browser for Kitty/Ghostty terminals using Kitty graphics protocol"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "trafilatura>=1.8.0",
    "html2text>=2024.2.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-httpserver>=1.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
webcat = "webcat.cli:main"

[tool.setuptools]
packages = ["webcat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
